backend/requirements.txt
fastapi
uvicorn

sqlalchemy
pydantic
dotenv
psycopg2-binary
python-dotenv